cleanup/requirements.txt
boto3
psycopg2
python-logstash==0.4.8